updated experience migration

+++ database/migrations/2019_08_15_193432_create_experiences_table.php
@@ -15,7 +15,7 @@ class CreateExperiencesTable extends Migration
     {
         Schema::create('experiences', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('user_id');
+            $table->unsignedInteger('user_id');
             $table->unsignedInteger('points');
             $table->timestamps();
         });
